requirements.txt
# Core Django Framework
Django==5.2.6
asgiref==3.9.2
sqlparse==0.5.3
tzdata==2025.2

# Django REST Framework for API development
djangorestframework==3.14.0

# JWT Authentication for REST API
djangorestframework-simplejwt==5.3.0
PyJWT==2.10.1

# CORS headers for Flutter app integration
django-cors-headers==4.3.1

# Swagger/OpenAPI documentation
drf-yasg==1.21.7
inflection==0.5.1
uritemplate==4.2.0
PyYAML==6.0.3
packaging==25.0

# API filtering
django-filter==24.3

# Firebase SDK and Google Cloud deps removed (unused)

# HTTP and networking
requests==2.32.5
urllib3==2.5.0
certifi==2025.8.3
charset-normalizer==3.4.3
idna==3.10

# HTTP/2 and async support
httpx==0.28.1
httpcore==1.0.9
h11==0.16.0
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
anyio==4.11.0
sniffio==1.3.1

# Image processing (for user profile images)
pillow==11.3.0

# PDF generation for invoices
reportlab==4.0.7

# Date utilities for month calculations
python-dateutil==2.8.2

# Excel file generation for reports
openpyxl==3.1.2

# Caching and utilities
CacheControl==0.14.3
cachetools==5.5.2
msgpack==1.1.1

# Cryptography
cryptography==43.0.1
cffi==1.17.1
pycparser==2.22

# Time zone support
pytz==2025.2

# Type hints support
typing-extensions==4.15.0

# Database (PostgreSQL support - if needed)
psycopg2-binary==2.9.10

# Development and deployment utilities (optional)
# python-decouple==3.8      # For environment variables management
# gunicorn==21.2.0          # For production deployment
# whitenoise==6.6.0         # For serving static files in production